requirements.txt
# Install Vertex AI LLM SDK
langchain==0.1.14 
langchain-google-vertexai==0.1.3 
typing_extensions==4.9.0

# Dependencies required by Unstructured PDF loader
poppler-utils
unstructured==0.12.4 
pdf2image==1.17.0 
pytesseract==0.3.10 
pdfminer.six==20221105
pillow-heif==0.15.0 
opencv-python==4.9.0.80 
unstructured-inference==0.7.24 
pikepdf==8.13.0 
pypdf==4.0.1
faiss-cpu
# tensorflow_hub==0.16.1 
# tensorflow_text==2.15.0
sentence-transformers
langchain-community 
#faiss-gpu==1.7.4
sentence_transformers > /dev/null
langchain_community
gpt4all
langchain
transformers
spacy
pdfplumber